In ares.py AresManager class created along with methods for retrieving data of given subject from ARES
Methods are
get_subject_by_ico - the core methods that is expected to be the most important. It returns subject data as a dictionary based on IČO from the ARES REST Api

get_subject_formal_description - uses above method to retrieve subject data, then picks the most important one and returns them in a format considered to be formal convention in written communication.