Separating ico validation into three sub-methods
Previously, the get_subject_by_ico() method took care in validating and manipulating the ico argument to be suited for ARES API.

However, this was not aligned with SOC and SRP principles.

ICO validation has been delegated into a separate method validate_ico(), which further delegates this task to _force_string() and _force_full_length() methods.

The code is now more modular.