creating AresApiClientManager class
this class will serve as a manager to interact with AresApiClient class.

AresApiClient class itself can be used to retrieve the data about the economic subject from ARES public register. It does not need any user interaction for this, if IČO parameter is passed by another programs.

However, with this manager, AresApiClient can be used to interact with the user through CLI to obtain subject formal description