improving ico validation - more robust tests
with the introduction of the interactive mode, a more robust input validation seems like a good idea. This should further enhance user experience as well as security.

Two additional validation methods were added to AresApiClient -> _check_if_is_digit() and _check_exceeding_allowed_length().

The validate_ico() method was restructured.

All the tests are now contained in a list of tests. The validate_ico() iterates through these, passing ico to each one of them, allowing it to be modified if needed. All the test methods return ico if passed (modified or not) or return None instead. Because of this logic, simple str.isdigit() was not used, since it would return boolean, but it was wrapped in a static method to return ico itself.

Failing any of the tests means validate_ico()  method immediately returns None. Passing all the tests returns ico, possibly modified to suit the ARES API.

Also, the AresApiClientManager now removes all possible whitespaces from user input, not just spaces.