handling the connection error in the interactive mode
In the interactive mode of the AresApiClientManager, the code trying to retrieve data from ARES API was put into the try-except block to be more robust against the loss of the internet connection.

The user now gets notified about the connection error, but the code does not crash and the user can continue after he resolves the connection issue.